third_party/blink/web_tests/http/tests/websocket/workers/connect-error-by-no-websocket-server-closeonerror-expected.txt

CONSOLE ERROR: WebSocket connection to 'ws://127.0.0.1:8892/' failed: Error in connection establishment: net::ERR_CONNECTION_REFUSED
Verify firing error event behaviors in Worker when close() is called from onerror.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/connect-error-by-no-websocket-server-closeonerror.js
[Worker] Test start (Port 8892)
PASS [Worker] onerror was called
PASS [Worker] onclose was called
PASS successfullyParsed is true

TEST COMPLETE
